requirements.txt
fastapi==0.111.1
uvicorn==0.30.1
pymysql==1.1.1
pynmea2==1.19.0
pyserial==3.5
# serial==0.0.97
python-dotenv==1.0.1
pika==1.3.2
gpiozero==2.0.1
smbus==1.1.post2